+++ changelog.md
@@ -11,6 +11,7 @@
 * Fixed issues with `setSrc` method duplicating events on the native renderers @rafa8626
 * Removed code to load media when playing considered unnecessary @rafa8626
 * Added mechanism in `YouTube` renderer to add poster based on its Image API using `youtube`'s `imageQuality` parameter @rafa8626
+* Fixed typo on `setFill` mode and fixed issue with poster when using iOS @rafa8626
 
 *4.2.1/4.2.2 (2017/06/28)*